Refactor reports menu
