--- Jolly-Lights-Cinema-Group/BusinessLogic/GlobalHandlers/ReportsHandler.cs
@@ -1,79 +1,0 @@
-using JollyLightsCinemaGroup.DataAccess;
-
-namespace Jolly_Lights_Cinema_Group
-{
-    public static class ReportsHandler
-    {
-        public static void ManageReports()
-        {
-            bool inManageReportsMenu = true;
-            ReportsMenu reportsMenu = new();
-            Console.Clear();
-
-            while (inManageReportsMenu)
-            {
-                int userChoice = reportsMenu.Run();
-                inManageReportsMenu = HandleManageReportsChoice(userChoice);
-                Console.Clear();
-            }
-        }
-        private static bool HandleManageReportsChoice(int choice)
-        {
-            switch (choice)
-            {
-                case 0:
-                    AnnualEarnings();
-                    return true;
-                case 1:
-                    return false;
-                default:
-                    Console.WriteLine("Invalid selection.");
-                    return true;
-            }
-        }
-
-        private static void AnnualEarnings()
-        {
-            Console.Clear();
-
-            CustomerOrderRepository customerOrderRepository = new();
-            List<int> years = customerOrderRepository.GetAvailableYears();
-
-            if (years.Count == 0)
-            {
-                Console.WriteLine("No years available yet.");
-                Console.ReadKey();
-                return;
-            }
-
-            string[] yearOptions = years.Select(year => year.ToString()).Append("Cancel").ToArray();
-            Menu yearMenu = new("Select a year:", yearOptions);
-            int selectedIndex = yearMenu.Run();
-
-            if (selectedIndex == yearOptions.Length - 1)
-            {
-                Console.WriteLine("Cancelled.");
-                return;
-            }
-
-            int selectedYear = years[selectedIndex];
-
-            CustomerOrderService customerOrderService = new();
-            double grossIncome = customerOrderService.GetGrossAnualEarnings(selectedYear);
-            double netIncome = customerOrderService.GetNetAnualEarnings(selectedYear);
-
-            Console.Clear();
-
-            Console.WriteLine($"Year: {selectedYear}");
-            Console.WriteLine($"----------------------------------");
-            Console.WriteLine($"Gross: €{grossIncome}");
-            Console.WriteLine($"Tax: €{Math.Round(grossIncome - netIncome, 2)}");
-            Console.WriteLine($"----------------------------------");
-            Console.WriteLine($"Net: €{netIncome}");
-
-            Console.WriteLine("\nPress any key to continue.");
-            Console.ReadKey();
-        }
-    }
-
-}
